feat(backend): permite media estrella en las reseñas
Review.rating pasa de Integer (1-5) a BigDecimal, en pasos de 0.5
(0.5 a 5.0). Nueva anotación de validación @HalfStep (con su
ConstraintValidator) para rechazar valores como 3.3 en el propio
DTO, junto a @DecimalMin/@DecimalMax para el rango.

Verificado en local: 3.5 se guarda bien, 3.3 se rechaza por el paso,
5.5 se rechaza por el rango. mvn test en verde.

+++ src/main/java/com/shelfy/review/dto/ReviewResponse.java
@@ -1,10 +1,11 @@
 package com.shelfy.review.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record ReviewResponse(
         Long id,
-        Integer rating,
+        BigDecimal rating,
         String text,
         Instant createdAt,
         Long bookId,
